Ensure script runs on all versions of the egap schema (#9470)
## Purpose

The egap migration script only got version 3 of the EGAP schema, so the 1536 EGAP registrations with version 2 of the schema wouldn't be migrated. This fixes that. 

## Changes

1. Loop over all of the versions of the schema
2. Assume the `_id` of the EGAP registry will be 'egap' (because it will be).
3. Rename `epag_provider` to `egap_provider`

## QA Notes

This will be tested on Staging2 as part of branded registries submission. It shouldn't be any more problematic than it would have been otherwise.

## Documentation

N/A

## Side Effects

None.

## Ticket

N/A