Fixing copyrights in the footers [ENG-2434] (#9554)
## Purpose

In a few of our email templates, the copyright year is hardcoded as 2018 in the footer.

## Changes

Make the emails use dynamic dates

## QA Notes

Please make verification statements inspired by your code and what your code touches.
- Verify the email is now 2020 in emails.
- Verify

What are the areas of risk? N/A

Any concerns/considerations/questions that development raised? N/A

## Documentation

N/A

## Side Effects

N/A
## Ticket

https://openscience.atlassian.net/browse/ENG-2434