Ensure withdrawal justification is saved as comment for withdrawal action [ENG-2453] (#9568)
## Purpose

The API needs to return the withdrawal justification as the comment in the withdrawal action. Its not doing this right now

## Changes

Ensure the registration model saves the withdrawal justification as the action comment.

## QA Notes

Please make verification statements inspired by your code and what your code touches.
- Verify the API returns the withdrawal justification made by the registration admin in the API
- Verify

What are the areas of risk? N/A

Any concerns/considerations/questions that development raised? N/A

## Documentation

N/A

## Side Effects

N/A
## Ticket

https://openscience.atlassian.net/browse/ENG-2453