[CR][ENG-5997] merge develop into b-and-i branch (#10691)
## Purpose

Bring `feature/b-and-i-24-14` up to date with `develop` (and the upgrade to python v3.12)

## Changes

* Some small conflict resolutions
* One change to make a `.format()` string an f-string

## QA Notes

This may merit a retesting, but I have reviewed the changes in the branch and they all seem pretty compatible with the py3.12 updates. 

## Documentation

No documentation changes needed.

## Side Effects

Devs will need to update their envs for the new py3.12 base.  That'll involve pulling images, rebuilding requirements, and maybe something else?

## Ticket

https://openscience.atlassian.net/browse/ENG-5997