ENG-5864 BE disable creator self removal and permission change before preprint is not submitted (#10978)
## Purpose

preventing the Creator of the preprint, while it’s a draft from changing their permissions or removing themselves as a contributor. 

## Changes

updating back end patch and delete views with restriction conditions

## QA Notes

Front end behaviour does not look the same in two different contributor update requests while creating a preprint. 

On the first one it was created a request to back end side and the answer to front end was returned (it is needed to investigate why it is not returned the error message that is on backend side). 

On second attempt a request for backend is not done (may see it on chrome dev tools on the video below).

it was tested on branch feature/b-and-i-25-01 (EMBER / OSF)


## Ticket

https://openscience.atlassian.net/browse/ENG-5864?focusedCommentId=83086