[ENG-5668] fixed ham functionality (#10977)
## Purpose

After creation of a registration of a spammed and hammed project and rejecting it as a moderator, we cannot ham project's creator

## Changes

When we clone logs to a registration, spam logs have `was_public = True` in `params` that is inaccurate as registration is private until all approvals and embargo. When it's True we run Node.SpamOverrideMixin.confirm_ham.undelete.set_privacy sequence of methods when ham a user that causes an error after rejecting the registration in moderation.

## Ticket

https://openscience.atlassian.net/browse/ENG-5668