ENG-6987 enable file by guid searching via osf admin (#10985)
## Purpose
As with nodes, preprints, and users, a lookup is available for files.  The admin detail page includes all available metadata including contributors from the node, the node title and link, version information, and file name.

## Changes
Adding in osf admin dashboard a page for searching files by guid and for showing the response

## Ticket
https://openscience.atlassian.net/browse/ENG-6987