[ENG-4255] update to public true on ham just the objects that were public before were spammed (#10970)
## Purpose
When a user is is designated as Spam, and then confirmed as Ham, sometimes their preprints remain set to private.
It is needed to update to public true on ham just the objects that were public before were spammed.

## Changes
It is needed to was_public parameter from the latest spammed log that was after public/private changings 

## Ticket
https://openscience.atlassian.net/browse/ENG-4255