[ENG-7460] Unpublished preprints in pre-moderation return 404 (#11010)
## Purpose
The current `get_preprint` method filters preprints by `is_published = True` that is incorrect in case `pre-moderation` is enabled as in this case preprints have `is_published = False` but still should be visible by users

## Changes
Fixed queryset, added tests

## Ticket
https://openscience.atlassian.net/browse/ENG-7460

##Notes
If it's never been approved for pre-moderation, then it should 404 for non-contributors, whether it was withdrawn or not. If you are a contributor, it should show for you in the proper state.

If it was approved for pre-moderation then withdrawn, it should show as withdrawn for everyone.