[ENG-8048] Remove caching to avoid incorrect results for ascendants (#11169)
## Purpose
When user adds a component to a node that has addons with files and registers it, the node doesn't have any archived files. It happens because the child nodes are the first ones to be registered and if a child doesn't have any addons, this empty list would be saved in `request.gv_addons` attribute that is reused for all other parents and children to avoid making redundant requests to GV. However each child and parent may have its own collection of addons.

## Changes
Removed caching

## QA Notes
Would be nice to test what happens when a child node has addons with files, but its parent hasn't. Will the parent inherit archived addons files or not?

## Ticket
https://openscience.atlassian.net/browse/ENG-8048