[ENG-7929] Ability to move registrations to draft state (#11153)
## Purpose
Admins should be able to revert registration to draft state for the support purposes

## Changes
Added functionality to revert registration back to draft state
Added a button in admin to revert a registration
Added tests

## Notes:
Currently we are discussing with Mark what to do with a registration after admin reverts it, so this ticket cannot be merged. The issue is that when admin reverts registration, it's still displayed for admin, just with deletion date and when user re-registers a draft, the system creates a new registration and the draft is linked to this newest version and admin sees both registrations. So if admin tries to revert the previous version, he'll get an error because the version doesn't have the linked draft. What I've suggested:
1. Fully delete previous registered version so that you won't be able to see it in admin. In this way if the registered version had guid 123ab and user registered the restored draft again, the draft would have another registration with another guid, like 999ss . Basically for user it'll be the same registration, just with another url to access it
2. Add another (or the same) hint for Revert to Draft button that you can't revert this previous version (the easiest solution)
3. One more solution, that I think will be the most complicated between these three, is that we can try to save the guid of an original registered version and whenever admin reverts registration and user registers it again from the same draft, we assign the saved guid to a new registration

**DECISION**:
Option 2 is the most reliable. In case we want #1 or #3 behavior, a new ticket will be created

## Ticket
https://openscience.atlassian.net/browse/ENG-7929