added a route to download node metadata (#11215)
## Purpose
So right now, our firewall rules aren't advanced enough to be able to let requests from `/<guid>/metadata` go through to osf on staging4. Instead, we have to start with a stable string, and do the dynamic part afterward, hence doing `/metadata/<guid>/`  which we can match with `/metadata/*` in the firewall.

## Changes
Added a corresponding route

## Ticket
https://openscience.atlassian.net/browse/ENG-8261