[ENG-8246] Fixed deletion of maintenance alerts in admin (#11226)
What

When you try to delete the active maintenance banner from the admin app, it 502s. The traceback says that there is no url to redirect to and to provide a success url. 

Acceptance Criteria

A user will be able to add a maintenance banner and then delete it. When the banner is added, it will show on the OSF dashboard page (among other places). When deleted, it will no longer show.